refactor to register sql provider only once
