+++ src/designer/elsa-workflows-studio/package.json
@@ -35,6 +35,7 @@
   "dependencies": {
     "@microsoft/signalr": "^6.0.3",
     "ajv": "^8.10.0",
+    "async-mutex": "^0.3.2",
     "axios": "^0.26.1",
     "axios-middleware": "^0.3.1",
     "cronstrue": "^1.125.0",
