+++ src/designer/elsa-workflows-studio/src/components/controls/elsa-monaco/elsa-monaco-utils.ts
@@ -1,4 +1,5 @@
-﻿const win = window as any;
+﻿import { Mutex } from "async-mutex";
+const win = window as any;
 const require = win.require;
 
 export interface Monaco {
@@ -8,12 +9,26 @@ export interface Monaco {
   Uri: any;
 }
 
+export interface EditorVariable {
+  variableName: string;
+  type: string;
+}
+
+export var editorVariables : Array<EditorVariable> = [];
+
 let isInitialized : boolean;
+const mutex = new Mutex();
 
-export function initializeMonacoWorker(libPath?: string): Promise<Monaco> {
 
-  if (isInitialized)
+
+export async function initializeMonacoWorker(libPath?: string): Promise<Monaco> {
+  
+  const release = await mutex.acquire();
+  
+  if (isInitialized) {
+    release();
     return win.monaco;
+  }
 
   const origin = document.location.origin;
   const baseUrl = libPath.startsWith('http') ? libPath : `${origin}/${libPath}`;
@@ -31,7 +46,68 @@ export function initializeMonacoWorker(libPath?: string): Promise<Monaco> {
   return new Promise(resolve => {
     require(["vs/editor/editor.main"], () => {
       isInitialized = true;
+      release();
+      registerLiquid(win.monaco);
+      registerSQL(win.monaco);
       resolve(win.monaco);
     });
   });
+
+
+  function registerLiquid(monaco: any) {
+    monaco.languages.register({ id: 'liquid' });
+
+    monaco.languages.registerCompletionItemProvider('liquid', {
+      provideCompletionItems: () => {
+        const autocompleteProviderItems = [];
+        const keywords = ['assign', 'capture', 'endcapture', 'increment', 'decrement',
+          'if', 'else', 'elsif', 'endif', 'for', 'endfor', 'break',
+          'continue', 'limit', 'offset', 'range', 'reversed', 'cols',
+          'case', 'endcase', 'when', 'block', 'endblock', 'true', 'false',
+          'in', 'unless', 'endunless', 'cycle', 'tablerow', 'endtablerow',
+          'contains', 'startswith', 'endswith', 'comment', 'endcomment',
+          'raw', 'endraw', 'editable', 'endentitylist', 'endentityview', 'endinclude',
+          'endmarker', 'entitylist', 'entityview', 'forloop', 'image', 'include',
+          'marker', 'outputcache', 'plugin', 'style', 'text', 'widget',
+          'abs', 'append', 'at_least', 'at_most', 'capitalize', 'ceil', 'compact',
+          'concat', 'date', 'default', 'divided_by', 'downcase', 'escape',
+          'escape_once', 'first', 'floor', 'join', 'last', 'lstrip', 'map',
+          'minus', 'modulo', 'newline_to_br', 'plus', 'prepend', 'remove',
+          'remove_first', 'replace', 'replace_first', 'reverse', 'round',
+          'rstrip', 'size', 'slice', 'sort', 'sort_natural', 'split', 'strip',
+          'strip_html', 'strip_newlines', 'times', 'truncate', 'truncatewords',
+          'uniq', 'upcase', 'url_decode', 'url_encode'];
+
+        for (let i = 0; i < keywords.length; i++) {
+          autocompleteProviderItems.push({ 'label': keywords[i], kind: monaco.languages.CompletionItemKind.Keyword });
+        }
+
+        return { suggestions: autocompleteProviderItems };
+      }
+    });
+  }
+
+  function registerSQL(monaco:any) {
+
+    monaco.languages.registerCompletionItemProvider('sql', {
+      triggerCharacters: ["@"],
+      provideCompletionItems: (model, position) => {
+
+        const word = model.getWordUntilPosition(position)
+
+        const autocompleteProviderItems = [];
+          for(const varible of editorVariables) {
+            autocompleteProviderItems.push({
+              label: `${varible.variableName}: ${varible.type}`,
+              kind: monaco.languages.CompletionItemKind.Keyword,
+              insertText: varible.variableName
+            });
+          }
+
+        return { suggestions: autocompleteProviderItems };
+      }
+    });
+
+  }
 }
+
